Move publishing actions to a mixin
